Add files via upload
CPU Process Scheduling project aims to do some function like(add, delete and count) on it by using stack and queue. and it also has binary tree to return the value by preorder and this class is separated from the other classes that working on the CPU process. and the class of the binary tree the user should give the number of nodes then give each node its value.